+++ src/braket/circuits/gates.py
@@ -19,7 +19,6 @@ from sympy import Float
 import braket.ir.jaqcd as ir
 from braket.circuits import circuit
 from braket.circuits.angled_gate import AngledGate
-from braket.circuits.free_parameter import FreeParameter
 from braket.circuits.free_parameter_expression import FreeParameterExpression
 from braket.circuits.gate import Gate
 from braket.circuits.instruction import Instruction
@@ -539,10 +538,10 @@ class Rx(AngledGate):
     """X-axis rotation gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: Union[FreeParameter, float]):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -568,27 +567,18 @@ class Rx(AngledGate):
         return 1
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.Rx: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
     @circuit.subroutine(register=True)
-    def rx(target: QubitInput, angle: Union[FreeParameter, float]) -> Iterable[Instruction]:
+    def rx(
+        target: QubitInput, angle: Union[FreeParameterExpression, float]
+    ) -> Iterable[Instruction]:
         """Registers this function into the circuit class.
 
         Args:
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): Angle in radians.
+            angle (Union[FreeParameterExpression, float]): Angle in radians.
 
         Returns:
             Iterable[Instruction]: Rx instruction.
@@ -606,10 +596,10 @@ class Ry(AngledGate):
     """Y-axis rotation gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: Union[FreeParameter, float]):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -635,27 +625,18 @@ class Ry(AngledGate):
         return 1
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.Ry: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
     @circuit.subroutine(register=True)
-    def ry(target: QubitInput, angle: Union[FreeParameter, float]) -> Iterable[Instruction]:
+    def ry(
+        target: QubitInput, angle: Union[FreeParameterExpression, float]
+    ) -> Iterable[Instruction]:
         """Registers this function into the circuit class.
 
         Args:
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): Angle in radians.
+            angle (Union[FreeParameterExpression, float]): Angle in radians.
 
         Returns:
             Iterable[Instruction]: Ry instruction.
@@ -673,10 +654,10 @@ class Rz(AngledGate):
     """Z-axis rotation gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: Union[FreeParameter, float]):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -698,17 +679,6 @@ class Rz(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.Rz: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -717,12 +687,14 @@ class Rz(AngledGate):
 
     @staticmethod
     @circuit.subroutine(register=True)
-    def rz(target: QubitInput, angle: Union[FreeParameter, float]) -> Iterable[Instruction]:
+    def rz(
+        target: QubitInput, angle: Union[FreeParameterExpression, float]
+    ) -> Iterable[Instruction]:
         """Registers this function into the circuit class.
 
         Args:
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Iterable[Instruction]: Rz instruction.
@@ -740,10 +712,10 @@ class PhaseShift(AngledGate):
     """Phase shift gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -757,17 +729,6 @@ class PhaseShift(AngledGate):
         return np.array([[1.0, 0.0], [0.0, np.exp(1j * self.angle)]], dtype=complex)
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.PhaseShift: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -776,12 +737,14 @@ class PhaseShift(AngledGate):
 
     @staticmethod
     @circuit.subroutine(register=True)
-    def phaseshift(target: QubitInput, angle: Union[FreeParameter, float]) -> Iterable[Instruction]:
+    def phaseshift(
+        target: QubitInput, angle: Union[FreeParameterExpression, float]
+    ) -> Iterable[Instruction]:
         """Registers this function into the circuit class.
 
         Args:
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Iterable[Instruction]: PhaseShift instruction.
@@ -960,10 +923,10 @@ class PSwap(AngledGate):
     """PSwap gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -988,17 +951,6 @@ class PSwap(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.PSwap: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1013,7 +965,7 @@ class PSwap(AngledGate):
         Args:
             target1 (Qubit or int): Target qubit 1 index.
             target2 (Qubit or int): Target qubit 2 index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: PSwap instruction.
@@ -1033,10 +985,10 @@ class XY(AngledGate):
     Reference: https://arxiv.org/abs/1912.04424v1
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1070,17 +1022,6 @@ class XY(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.XY: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1090,14 +1031,14 @@ class XY(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def xy(
-        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameter, float]
+        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             target1 (Qubit or int): Target qubit 1 index.
             target2 (Qubit or int): Target qubit 2 index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: XY instruction.
@@ -1115,10 +1056,10 @@ class CPhaseShift(AngledGate):
     """Controlled phase shift gate.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1139,17 +1080,6 @@ class CPhaseShift(AngledGate):
         return np.diag([1.0, 1.0, 1.0, np.exp(1j * self.angle)])
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.CPhaseShift: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1159,14 +1089,14 @@ class CPhaseShift(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def cphaseshift(
-        control: QubitInput, target: QubitInput, angle: Union[FreeParameter, float]
+        control: QubitInput, target: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             control (Qubit or int): Control qubit index.
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: CPhaseShift instruction.
@@ -1184,10 +1114,10 @@ class CPhaseShift00(AngledGate):
     """Controlled phase shift gate for phasing the \\|00> state.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1208,17 +1138,6 @@ class CPhaseShift00(AngledGate):
         return np.diag([np.exp(1j * self.angle), 1.0, 1.0, 1.0])
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.CPhaseShift00: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1228,14 +1147,14 @@ class CPhaseShift00(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def cphaseshift00(
-        control: QubitInput, target: QubitInput, angle: Union[FreeParameter, float]
+        control: QubitInput, target: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             control (Qubit or int): Control qubit index.
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: CPhaseShift00 instruction.
@@ -1253,10 +1172,10 @@ class CPhaseShift01(AngledGate):
     """Controlled phase shift gate for phasing the \\|01> state.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1277,17 +1196,6 @@ class CPhaseShift01(AngledGate):
         return np.diag([1.0, np.exp(1j * self.angle), 1.0, 1.0])
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.CPhaseShift01: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1297,14 +1205,14 @@ class CPhaseShift01(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def cphaseshift01(
-        control: QubitInput, target: QubitInput, angle: Union[FreeParameter, float]
+        control: QubitInput, target: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             control (Qubit or int): Control qubit index.
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: CPhaseShift01 instruction.
@@ -1322,10 +1230,10 @@ class CPhaseShift10(AngledGate):
     """Controlled phase shift gate for phasing the \\|10> state.
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1339,17 +1247,6 @@ class CPhaseShift10(AngledGate):
         return np.diag([1.0, 1.0, np.exp(1j * self.angle), 1.0])
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.CPhaseShift10: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1359,14 +1256,14 @@ class CPhaseShift10(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def cphaseshift10(
-        control: QubitInput, target: QubitInput, angle: Union[FreeParameter, float]
+        control: QubitInput, target: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             control (Qubit or int): Control qubit index.
             target (Qubit or int): Target qubit index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: CPhaseShift10 instruction.
@@ -1597,10 +1494,10 @@ class XX(AngledGate):
     Reference: https://arxiv.org/abs/1707.06356
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1634,17 +1531,6 @@ class XX(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.XX: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1654,14 +1540,14 @@ class XX(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def xx(
-        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameter, float]
+        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             target1 (Qubit or int): Target qubit 1 index.
             target2 (Qubit or int): Target qubit 2 index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: XX instruction.
@@ -1681,10 +1567,10 @@ class YY(AngledGate):
     Reference: https://arxiv.org/abs/1707.06356
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1718,17 +1604,6 @@ class YY(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.YY: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1738,14 +1613,14 @@ class YY(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def yy(
-        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameter, float]
+        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             target1 (Qubit or int): Target qubit 1 index.
             target2 (Qubit or int): Target qubit 2 index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: YY instruction.
@@ -1765,10 +1640,10 @@ class ZZ(AngledGate):
     Reference: https://arxiv.org/abs/1707.06356
 
     Args:
-        angle (Union[FreeParameter, float]): angle in radians.
+        angle (Union[FreeParameterExpression, float]): angle in radians.
     """
 
-    def __init__(self, angle: float):
+    def __init__(self, angle: Union[FreeParameterExpression, float]):
         super().__init__(
             angle=angle,
             qubit_count=None,
@@ -1800,17 +1675,6 @@ class ZZ(AngledGate):
         )
 
     def bind_values(self, **kwargs):
-        """
-        Takes in parameters and attempts to assign them to values.
-
-        Args:
-            **kwargs: The parameters that are being assigned.
-
-        Returns:
-            Gate.ZZ: A new Gate of the same type with the requested
-            parameters bound.
-
-        """
         return get_angle(self, **kwargs)
 
     @staticmethod
@@ -1820,14 +1684,14 @@ class ZZ(AngledGate):
     @staticmethod
     @circuit.subroutine(register=True)
     def zz(
-        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameter, float]
+        target1: QubitInput, target2: QubitInput, angle: Union[FreeParameterExpression, float]
     ) -> Instruction:
         """Registers this function into the circuit class.
 
         Args:
             target1 (Qubit or int): Target qubit 1 index.
             target2 (Qubit or int): Target qubit 2 index.
-            angle (Union[FreeParameter, float]): angle in radians.
+            angle (Union[FreeParameterExpression, float]): angle in radians.
 
         Returns:
             Instruction: ZZ instruction.
@@ -2036,13 +1900,13 @@ class Unitary(Gate):
 Gate.register_gate(Unitary)
 
 
-def angled_ascii_characters(gate: str, angle: Union[FreeParameter, float]) -> str:
+def angled_ascii_characters(gate: str, angle: Union[FreeParameterExpression, float]) -> str:
     """
     Generates a formatted ascii representation of an angled gate.
 
     Args:
         gate (str): The name of the gate.
-        angle (Union[FreeParameter, float]): The angle for the gate.
+        angle (Union[FreeParameterExpression, float]): The angle for the gate.
 
     Returns:
         str: Returns the ascii representation for an angled gate.
@@ -2061,7 +1925,6 @@ def get_angle(self, **kwargs):
 
     Returns:
         A new gate of the type of the AngledGate originally used with all angles updated.
-
     """
     new_angle = (
         self.angle.subs(kwargs) if isinstance(self.angle, FreeParameterExpression) else self.angle